Combine Dockerfile update with Build and Deploy workflow
Making a commit from the Update base image is not triggering the build and
deploy workflow. Assumption is this is a GitHub protection to avoid the risk of
recursion.

Solution is to combine the steps to update the Dockerfile into the Build and
Deploy workflow, just running them conditionally if triggered from event
dispatch.

I'm not able to split it into a separate job because the job would need to be
run unconditionally.