Replace typographic quotation marks with neutral
As these files will be rendered in the web app using Kramdown, complete
with its "smart quotes" plugin, we're able to store the documents with
regular quote marks, allowing for easier editing in text editors while
maintaining consistency.

The neutral quotes will be automatically switched back to typographic
ones when the document is rendered.