Fix button classes for Govspeak context
As these guidance pages are picking up their style info from Govspeak
rather than plain old GOV.UK Design System, we need to tweak them a bit
by adding `gem-c-button` in. There's a full explanation of the styling
options here, but as we're not using the full gem we need the generated
markup not the source markdown.

https://govspeak-preview.herokuapp.com/guide