Separate return to teaching content to partials
This allows a CTA to be placed after the intro and for us to reorder and
organise things more freely. Currently the right hand column only
supports one 'type' of content so we're going with jump links, the CTA
has been left in a commented out fashion as it will soon be enabled.