Move and rename story list pages
These pages are unusual in that they are at `/section/index.md` and yet we
serve them at `/section` instead of `/section/index` (they used a custom
controller in the app).

We have changed the app to support rendering these at both `/section/index.md`
and `/section.md` so that we can rename and move them without changing their
URLs. Once moved we will be able to pull out some temporary code in the app
that supports this.